Improve highlighting detection
Currently, xtext considers any mouse motion after a click to be the
start of a highlight. GTK seems to be extraordinarily sensitive with
this motion event, and fires it even if the mouse has imperceptibly
moved. This means that what feels like a single click to a user is
detected by xtext as a highlight.

This commit makes things slightly better by checking to see if the mouse
has moved at least five pixels before starting to highlight text. This
makes the single click event fire more reliably.